(hotfix: redis-rollback)
제가 작업했던 레디스를 모두 롤백 했습니다.
너무 많은 이슈가 생겨서 일단은 전으로 돌렸습니다.
member 패키지에 있던 cache를 configuration으로 옮기고 수정했습니다.
