+++ build.gradle
@@ -42,7 +42,7 @@ dependencies {
     implementation 'com.github.gavlyukovskiy:p6spy-spring-boot-starter:1.9.0'
     
     // json 파싱 Jackson
-    implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.16.0'
+    implementation group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version: '2.15.3'
 
     // Websocket
     implementation 'org.springframework.boot:spring-boot-starter-websocket'
